# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 梗星球 MemeChatAI REST API
  version: 1.0.0
  description: |
    梗星球（MemeChatAI）REST API 接口契约，OpenAPI 3.1.0 机器可读版本。

    - 字段命名严格对齐 `db/schema.sql` v1.0.0（snake_case）。
    - 统一响应包装 `{code, data, message, request_id}`。
    - 鉴权：JWT Bearer（除标注为无需鉴权的接口）。
    - 异步 AI 任务：202 + job_id + 轮询/WebSocket。
    - M1 P0 接口对应 M1 Sprint（2026-07-08 ~ 2026-08-04）必交付范围；
      M2 接口对应第 2 个月交付；v1.5+ 接口在表格中列出但本文件给最小 schema。
    - 人类可读详设见 `docs/API-Spec.md`。
  contact:
    name: 后端架构 / API 设计
  license:
    name: Proprietary
  x-memestar-doc-version: v1.0.0
  x-memestar-related:
    prd: docs/PRD.md v1.1.0
    tech: docs/TechnicalDesign.md v1.1.0
    m1-plan: docs/M1-Sprint-Plan.md v1.0.0
    schema: docs/db/schema.sql v1.0.0
servers:
  - url: https://api.dev.memestar.cn/api/v1
    description: dev 本地开发联调
  - url: https://api.staging.memestar.cn/api/v1
    description: staging 灰度前联调 + 邀请制灰度
  - url: https://api.memestar.cn/api/v1
    description: prod 生产（ICP 备案后启用）
...
security:
  - BearerAuth: []
paths:
...
  /videos:
    post:
      tags: [Video]
      summary: 视频生成提交（异步）
      description: |
        异步任务，返回 202 + job_id。
        - 配额：免费 1 次/周（需任务解锁）/ Pro 3 次/日
        - 模型：豆包 Seedance 2.0 mini（主力）/ 标准版（Pro 高端）/ SiliconFlow fallback / 图片+TTS 兜底
        - 时长：5 / 10 / 15 秒
        - 日预算熔断：超 ¥100 自动降级到图片+TTS，超 ¥200 暂停真视频
      operationId: submitVideo
      x-memestar-version: M2
      x-memestar-rate-limit: free 1/week, pro 3/day
      x-memestar-async: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoSubmitRequest'
      responses:
        '202':
          description: 已接受
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VideoJobAck'
        '400':
          description: 能量不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                energy_low:
                  value: { code: 6001, data: null, message: "能量不足", request_id: "..." }
        '429':
          description: 配额耗尽 / 熔断
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                quota_out:
                  value: { code: 3102, data: null, message: "视频配额耗尽", request_id: "..." }
  /videos/{id}/status:
    get:
      tags: [Video]
      summary: 视频生成状态查询
      description: |
        客户端每 3s 轮询，最长 180s 后切 Webhook 推送。
        status：generating / reviewing / published / rejected / timeout。
        is_fallback=true 表示已降级为图片+TTS+Ken Burns 兜底。
      operationId: getVideoStatus
      x-memestar-version: M2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 状态
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VideoStatus'

  # ============================================================
  # Meme
  # ============================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer Token。Header 格式：`Authorization: Bearer <access_token>`。
        access_token 有效期 2h，refresh_token 30d。

  # ============================================================
  # Parameters
  # ============================================================
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 写操作幂等键，24h 内相同 key 返回首次响应。
...
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0=成功，非 0=业务错误码
          example: 0
        data:
          description: 业务数据，错误时为 null
        message:
          type: string
          example: ok
        request_id:
          type: string
          description: 32 位 ULID，链路追踪
          example: 01HQK8X9J2RT7N1V4Y6Z3B5M6P
...
    VideoSubmitRequest:
      type: object
      required: [source_type, duration]
      properties:
        creation_id:
          type: string
          format: uuid
          nullable: true
        source_type:
          type: string
          enum: [text_to_video, image_to_video, fallback_image_tts]
        duration:
          type: integer
          enum: [5, 10, 15]
        voice_id:
          type: string
          enum: [funny, northeast, queen, robot]
        subtitle_text:
          type: string
    VideoJobAck:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        video_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
        poll_url:
          type: string
        is_fallback:
          type: boolean
        estimated_seconds:
          type: integer
    VideoStatus:
      type: object
      description: 对齐 meme_videos 表
      properties:
        video_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [generating, reviewing, published, rejected, timeout]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        is_fallback:
          type: boolean
        model_version:
          type: string
          example: seedance-2-mini
        file_url:
          type: string
          nullable: true
        cover_url:
          type: string
          nullable: true
        duration:
          type: integer
        voice_id:
          type: string
          nullable: true
        subtitle_text:
          type: string
          nullable: true
        ai_cost:
          type: number
          format: float
        error:
          type: string
          nullable: true

    # ---------- Meme ----------
...